:recycle: Update install/uninstall scripts.

+++ scripts/ubuntu20.04/uninstall.sh
@@ -4,69 +4,27 @@
 set -e
 
 ###########################
-# Prerequisites
-###########################
-
-# Git
-if dpkg --get-selections | grep 'git'
-then
-  sudo apt autoremove --purge -y 'git'
-fi
-if test -f ~/.gitconfig
-then
-  rm -f ~/.gitconfig
-fi
-
-###########################
-# Terminal
-###########################
-
-# Oh My Zsh
-if test -d ~/.oh-my-zsh
-then
-  echo 'y' | sh ~/.oh-my-zsh/tools/uninstall.sh
-fi
-if test -f ~/.shell.pre-oh-my-zsh
-then
-  rm -f ~/.shell.pre-oh-my-zsh
-fi
-
-# Set default shell to bash
-if ! grep "^${USER}" /etc/passwd | grep '/bin/bash'
-then
-  chsh -s /bin/bash
-fi
-
-# Zsh
-if dpkg --get-selections | grep 'zsh'
-then
-  sudo apt autoremove --purge -y zsh
-fi
-if ls ~/.zsh* 2> /dev/null
-then
-  rm -f ~/.zsh*
-fi
-if ls ~/.zcomp* 2> /dev/null
-then
-  rm -f ~/.zcomp*
-fi
-
-###########################
 # Browsers
 ###########################
 
 # Chromium
-killall -q chrome
+if pgrep 'chrome'
+then
+  killall chrome
+fi
 if sudo snap list | grep 'chromium'
 then
   sudo snap remove --purge 'chromium'
 fi
 
 # Firefox
-killall -q firefox
+if pgrep 'firefox'
+then
+  killall firefox
+fi
 if dpkg --get-selections | grep 'firefox'
 then
-  sudo apt autoremove --purge  'firefox'
+  sudo apt autoremove -y --purge  'firefox'
 fi
 if test -d ~/.mozilla
 then
@@ -78,14 +36,17 @@ fi
 ###########################
 
 # Thunderbird
-killall -q thunderbird
+if pgrep 'thunderbird'
+then
+  killall thunderbird
+fi
 if ls -l /etc/apt/sources.list.d/ubuntu-mozilla-daily-ubuntu-ppa* 2> /dev/null
 then
   sudo rm -f /etc/apt/sources.list.d/ubuntu-mozilla-daily-ubuntu-ppa*
 fi
 if dpkg --get-selections | grep 'thunderbird'
 then
-  sudo apt autoremove --purge 'thunderbird'
+  sudo apt autoremove -y --purge 'thunderbird'
 fi
 if test -d ~/.thunderbird
 then
@@ -97,7 +58,10 @@ fi
 ###########################
 
 # Bitwarden
-killall -q bitwarden
+if pgrep 'bitwarden'
+then
+  killall bitwarden
+fi
 if sudo snap list | grep 'bitwarden'
 then
   sudo snap remove --purge 'bitwarden'
@@ -108,14 +72,20 @@ fi
 ###########################
 
 # LibreOffce
-killall -q soffice.bin
+if pgrep 'soffice.bin'
+then
+  killall soffice.bin
+fi
 if sudo snap list | grep 'libreoffice'
 then
   sudo snap remove --purge 'libreoffice'
 fi
 
 # Scribus
-killall -q scribus
+if pgrep 'scribus'
+then
+  killall scribus
+fi
 if ls -l /etc/apt/sources.list.d/scribus-ubuntu-ppa* 2> /dev/null
 then
   sudo rm -f /etc/apt/sources.list.d/scribus-ubuntu-ppa*
@@ -129,13 +99,15 @@ then
   rm -rf ~/.config/scribus
 fi
 
-
 ###########################
 # File viewers
 ###########################
 
 # PDF Sudio
-killall -q java
+if pgrep 'java'
+then
+  killall java
+fi
 if test -f ~/PDFStudio_linux64.sh
 then
   rm -f ~/PDFStudio_linux64.sh
@@ -156,10 +128,16 @@ if test -f ~/.install4j
 then
   rm -f ~/.install4j
 fi
-
+if ls -l ~/.gnome/apps/install4j* 2> /dev/null
+then
+  rm -f ~/.gnome/apps/install4j*
+fi
 
 # VLC
-killall -q vlc
+if pgrep 'vlc'
+then
+  killall vlc
+fi
 if sudo snap list | grep 'vlc'
 then
   sudo snap remove --purge 'vlc'
@@ -170,14 +148,20 @@ fi
 ###########################
 
 # Krita
-killall krita
+if pgrep 'krita'
+then
+  killall krita
+fi
 if sudo snap list | grep 'krita'
 then
   sudo snap remove --purge 'krita'
 fi
 
 # Inkscape
-killall inkscape
+if pgrep 'inkscape'
+then
+  killall inkscape
+fi
 if sudo snap list | grep 'inkscape'
 then
   sudo snap remove --purge 'inkscape'
@@ -193,7 +177,10 @@ fi
 ###########################
 
 # OpenShot Video Editor
-killall -q openshot-qt
+if pgrep 'openshot-qt'
+then
+  killall openshot-qt
+fi
 if ls -l /etc/apt/sources.list.d/openshot_developers-ubuntu-ppa* 2> /dev/null
 then
   sudo rm -f /etc/apt/sources.list.d/openshot_developers-ubuntu-ppa*
@@ -208,7 +195,10 @@ then
 fi
 
 # Handbrake
-killall ghb
+if pgrep 'ghb'
+then
+  killall ghb
+fi
 if ls -l /etc/apt/sources.list.d/stebbins-ubuntu-handbrake-releases* 2> /dev/null
 then
   sudo rm -f /etc/apt/sources.list.d/stebbins-ubuntu-handbrake-releases*
@@ -227,7 +217,10 @@ fi
 ###########################
 
 # OBS Studio
-killall -q obs
+if pgrep 'obs'
+then
+  killall obs
+fi
 if ls -l /etc/apt/sources.list.d/obsproject-ubuntu-obs-studio* 2> /dev/null
 then
   sudo rm -f /etc/apt/sources.list.d/obsproject-ubuntu-obs-studio*
@@ -246,16 +239,15 @@ fi
 ###########################
 
 # VS Code
-sudo pkill code
+if pgrep 'code'
+then
+  killall code
+fi
 if ls -l /etc/apt/sources.list.d/vscode* 2> /dev/null
 then
   sudo rm -f /etc/apt/sources.list.d/vscode*
 fi
-if dpkg --get-selections | grep "^code$"
-then
-  sudo apt autoremove --purge -y 'code'
-fi
-if dpkg --get-selections | grep "^code$"
+if dpkg --get-selections | grep "^code"
 then
   sudo apt autoremove --purge -y 'code'
 fi
@@ -281,7 +273,10 @@ fi
 ###########################
 
 # Filezilla
-killall -q filezilla
+if pgrep 'filezilla'
+then
+  killall filezilla
+fi
 if dpkg --get-selections | grep 'filezilla'
 then
   sudo apt autoremove --purge -y 'filezilla'
@@ -296,7 +291,10 @@ fi
 ###########################
 
 # Docker
-sudo pkill docker
+if pgrep 'docker'
+then
+  sudo pkill docker
+fi
 if sudo snap list | grep 'docker'
 then
   sudo snap remove --purge 'docker'
@@ -305,7 +303,7 @@ if groups "$USER" | grep 'docker'
 then
   sudo gpasswd -d "$USER" docker
 fi
-if less /etc/group | grep 'docker'
+if cat /etc/group | grep 'docker'
 then
   sudo groupdel 'docker'
 fi
@@ -315,13 +313,20 @@ fi
 ###########################
 
 # DBeaver
-killall -q java
+if pgrep 'java'
+then
+  killall java
+fi
 if sudo snap list | grep 'dbeaver-ce'
 then
   sudo snap remove --purge 'dbeaver-ce'
 fi
 
 # Robot 3T
+if pgrep 'robo3t'
+then
+  killall robo3t
+fi
 if sudo snap list | grep 'robo3t-snap'
 then
   sudo snap remove --purge 'robo3t-snap'
@@ -336,3 +341,86 @@ if sudo snap list | grep 'node'
 then
   sudo snap remove --purge 'node'
 fi
+if test -d ~/.npm
+then
+  rm -rf ~/.npm
+fi
+if test -d ~/v8-compile-cache-0
+then
+  sudo rm -rf ~/v8-compile-cache-0
+fi
+
+###########################
+# Terminal
+###########################
+
+# Set default shell to bash
+if ! grep "^${USER}" /etc/passwd | grep '/bin/bash'
+then
+  chsh -s /bin/bash
+fi
+
+# Oh My Zsh
+if test -f ~/.shell.pre-oh-my-zsh
+then
+  rm -f ~/.shell.pre-oh-my-zsh
+fi
+if test -d ~/.oh-my-zsh
+then
+  echo 'y' | sh ~/.oh-my-zsh/tools/uninstall.sh
+fi
+
+# Zsh
+if dpkg --get-selections | grep 'zsh'
+then
+  sudo apt autoremove --purge -y zsh
+fi
+if ls ~/.zsh* 2> /dev/null
+then
+  rm -f ~/.zsh*
+fi
+if ls ~/.zcomp* 2> /dev/null
+then
+  rm -f ~/.zcomp*
+fi
+
+###########################
+# Miscellaneous
+###########################
+
+# Gnome
+if test -d ~/.gnome
+then
+  rm -rf ~/.gnome
+fi
+
+# Gnupg
+if test -d ~/.gnupg
+then
+  rm -rf ~/.gnupg
+fi
+
+# Gphoto
+if test -d ~/.gphoto
+then
+  rm -rf ~/.gphoto
+fi
+
+# Pki
+if test -d ~/.pki
+then
+  rm -rf ~/.pki
+fi
+
+# Swt
+if test -d ~/.swt
+then
+  rm -rf ~/.swt
+fi
+
+# Wget
+if test -d ~/.wget-hsts
+then
+  rm -rf ~/.wget-hsts
+fi
+
